fix: resolve cyclic imports by extracting get_csp_nonce_from_request to utils
Move get_csp_nonce_from_request from mcpgateway.main to mcpgateway.utils.csp_nonce
to break the import cycle: mcpgateway.main -> ... -> mcpgateway.version -> mcpgateway.main

- Create mcpgateway/utils/csp_nonce.py with the extracted function
- Update mcpgateway/main.py to import from new location
- Update mcpgateway/version.py to import from new location
- Update tests/unit/mcpgateway/test_main.py to import from new location
- Update .secrets.baseline (line number shifts from new file)
