+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": "(?x)( package-lock\\.json$ |Cargo\\.lock$ |uv\\.lock$ |go\\.sum$ |mcpgateway/sri_hashes\\.json$ )|^.secrets.baseline$",
     "lines": null
   },
-  "generated_at": "2026-05-09T09:16:12Z",
+  "generated_at": "2026-05-09T13:03:11Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -5282,7 +5282,7 @@
         "hashed_secret": "6947818ac409551f11fbaa78f0ea6391960aa5b8",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 528,
+        "line_number": 582,
         "type": "Secret Keyword",
         "verified_result": null
       }
